+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ---
 
 ### New
+* services.cfn.deploy_template supports input parameters
 * Enable toggling aborting CodeBuild phases on command execution failure
 * Support CodeBuild spec runtime-versions